✨ Feat : 페이지네이션 구현
게시물과 유저의 전체 조회에서 페이지네이션 기능을 구현했다.

+++ src/main/java/com/fluffygram/newsfeed/domain/board/service/BoardService.java
@@ -6,8 +6,10 @@ import com.fluffygram.newsfeed.domain.board.entity.Board;
 import com.fluffygram.newsfeed.domain.board.repository.BoardRepository;
 import com.fluffygram.newsfeed.domain.user.entity.User;
 import com.fluffygram.newsfeed.domain.user.repository.UserRepository;
-import jakarta.validation.constraints.NotBlank;
+import com.fluffygram.newsfeed.global.exception.BusinessException;
+import com.fluffygram.newsfeed.global.exception.ExceptionType;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -18,26 +20,27 @@ public class BoardService {
     private final UserRepository userRepository;
     private final BoardRepository boardRepository;
 
-    //--로그인 세션 사용하기
-
     //게시물 생성(저장)
-    public BoardResponseDto save(Long id, String title, String contents){
-        //사용자 id로 사용자 조회
-        User findUserById = userRepository.findByIdOrElseThrow(id);
+    public BoardResponseDto save(Long userId, String title, String contents, Long loginUserId) {
+        // 로그인한 사용자와 아이디(id) 일치 여부 확인
+        if(!loginUserId.equals(userId)) {
+            throw new BusinessException(ExceptionType.USER_NOT_MATCH);
+        }
 
-        Board board = new Board(id, title, contents, findUserById);
+        //사용자 id로 사용자 조회
+        User findUserById = userRepository.findByIdOrElseThrow(userId);
+
+        Board board = new Board(title, contents, findUserById);
 
         //게시물 저장
         Board saveBoard = boardRepository.save(board);
 
-//        --.toDto(saveBoard) 사용하기
-//        return new BoardResponseDto(saveBoard.getId(), saveBoard.getTitle(), saveBoard.getContents(), findUserById.getUserNickname(), saveBoard.getCreatedAt(), saveBoard.getModifiedAt());
         return BoardResponseDto.toDto(saveBoard);
     }
 
     //게시물 전체 List 조회
-    public List<BoardListResponseDto> findAllBoard() {
-        return boardRepository.findAll()
+    public List<BoardListResponseDto> findAllBoard(Pageable pageable) {
+        return boardRepository.findAllByOrderByCreatedAtDesc(pageable)
                 .stream()
                 .map(BoardListResponseDto::toDto)
                 .toList();
@@ -48,38 +51,37 @@ public class BoardService {
         //게시물 조회
         Board findBoard = boardRepository.findBoardByIdOrElseThrow(id);
 
-        //사용자 id로 사용자 조회 -> 사용자 id로 조회하고있지 않다
-//        User findUserById = userRepository.findByIdOrElseThrow(id);
-
-        return new BoardResponseDto(findBoard.getId(),findBoard.getTitle(),findBoard.getContents(), findBoard.getUser().getUserNickname(), findBoard.getCreatedAt(),findBoard.getModifiedAt());
+        return BoardResponseDto.toDto(findBoard);
     }
 
     //게시물 id로 특정 게시물 수정
-    public BoardResponseDto updateBoard(Long id, String title, String contents) {
+    public BoardResponseDto updateBoard(Long id, String title, String contents, Long loginUserId) {
         //게시물 조회
         Board findBoard = boardRepository.findBoardByIdOrElseThrow(id);
 
-//        //사용자 id로 사용자 조회
-//        User findUserById = userRepository.findByIdOrElseThrow(id);
-
-        //--entity로 옮기기
-        if(title != null){
-            findBoard.updateTitle(title);
+        // 로그인한 사용자와 아이디(id) 일치 여부 확인
+        if(!loginUserId.equals(findBoard.getUser().getId())) {
+            throw new BusinessException(ExceptionType.USER_NOT_MATCH);
         }
 
-        if(contents != null){
-            findBoard.updateContents(contents);
-        }
+        // 제목 및 내용 수정하기
+        findBoard = findBoard.updateBoard(title, contents);
 
         //수정된 게시물 저장하기
         Board saveBoard = boardRepository.save(findBoard);
-        return new BoardResponseDto(saveBoard.getId(), saveBoard.getTitle(), saveBoard.getContents(), findBoard.getUser().getUserNickname(), saveBoard.getCreatedAt(), saveBoard.getModifiedAt());
+
+        return BoardResponseDto.toDto(saveBoard);
     }
     
     //게시물 id로 특정 게시물 삭제
-    public void deleteBoard(Long id) {
+    public void deleteBoard(Long id, Long loginUserId) {
         //게시물 조회
         Board findBoard = boardRepository.findBoardByIdOrElseThrow(id);
+
+        // 로그인한 사용자와 아이디(id) 일치 여부 확인
+        if(!loginUserId.equals(findBoard.getUser().getId())) {
+            throw new BusinessException(ExceptionType.USER_NOT_MATCH);
+        }
 
         boardRepository.delete(findBoard);
     }
